fix: show that a ruleset is modfied (#768)
Currently it is not possible to see if the selected ruleset has been modified. 
Furthermore, the button on the settings sub pages says reset to default
which was not what happened. This commit shows in the list if the ruleset has been
modified, and provides colored borders around the relevant settings and if they have 
been modified. It also shows the text "close without saving" rather than "reset to
default" on the sub pages. Refactored and moved settings related things
around.

Co-authored-by: marvin9257 <72580196+marvin9257@users.noreply.github.com>