feat: add a deck plan hyperlink for ship (#836)
This enhancement allows GM's to drag and drop a scene onto a ship sheet.  It is meant to be a link to the ship's deck plan (map of the ship) so that one jump to that scene quickly.  To use, drag the scene to the ship sheet.  A link should appear under the ship image that says "Deck Plan" in English.  Next to it is a trash can to delete the link, if desired.