feat: add third alternate characteristic (#1536)
This release adds a third alternative characteristic to Traveller, Animal and Robot sheets when using the "All" setting for "Select characteristics to display". So now, one can have up to ten characteristics.  For the Traveller sheet, it is recommended not to use the table display if using all ten characteristics.  Note that this update required some refactoring changes.  So, please test before using for real play.  

Also included are some localization updates from @ForjaSalvaje.